+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
+## [1.1.0] - 2023-11-08
+### Added
+- Added ability to search components for replacement in other C# assemblies
+- Added ability to ignore some base classes
+- Added a new page in "Project Settings/Component Replacer Settings" to specify ignored assemblies and base classes
+
 ## [1.0.0] - 2022-02-10
 ### Added
 - Initial Commit with all files